+++ lifelines/fitters/spline_fitter.py
@@ -5,22 +5,7 @@ from lifelines.utils.safe_exp import safe_exp
 from lifelines import utils
 
 
-class SplineFitterMixin:
-    _scipy_fit_method = "SLSQP"
-    _scipy_fit_options = {"ftol": 1e-10}
-
-    @staticmethod
-    def relu(x):
-        return np.maximum(0, x)
-
-    def basis(self, x, knot, min_knot, max_knot):
-        lambda_ = (max_knot - knot) / (max_knot - min_knot)
-        return self.relu(x - knot) ** 3 - (
-            lambda_ * self.relu(x - min_knot) ** 3 + (1 - lambda_) * self.relu(x - max_knot) ** 3
-        )
-
-
-class SplineFitter(SplineFitterMixin, KnownModelParametricUnivariateFitter):
+class SplineFitter(utils.SplineFitterMixin, KnownModelParametricUnivariateFitter):
     r"""
     Model the cumulative hazard using cubic splines. This offers great flexibility and smoothness of the cumulative hazard.